tests almost working
