+++ src/mechaphlowers/core/geometry/points.py
@@ -1,249 +1,228 @@
-# Copyright (c) 2025, RTE (http://www.rte-france.com)
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-# SPDX-License-Identifier: MPL-2.0
+from typing import Self, Tuple
 
 import numpy as np
 
-from mechaphlowers.core.geometry.rotation import (
-    rotation_quaternion,
-    rotation_quaternion_same_axis,
+from mechaphlowers.config import options as cfg
+from mechaphlowers.core.geometry.line_angles import (
+    get_insulator_layer,
+    get_supports_layer,
 )
+from mechaphlowers.core.geometry.references import (
+    CablePlane,
+    cable_to_beta_plane,
+    cable_to_crossarm_frame,
+    get_supports_coords,
+    translate_cable_to_support,
+)
+from mechaphlowers.core.models.cable.span import Span
 
 
-class Frame:
-    def __init__(
-        self,
-        origin: np.ndarray = np.full((4, 3), np.zeros(3)),
-        axis: np.ndarray = np.full((4, 3, 3), np.eye(3)),
-    ):
-        self.origin: np.ndarray = origin
-        #    [[x_0, y_0, z_0], #layer 0
-        #     [x_1, y_1, z_1], #layer 1
-        #     ...
-        #    ]
-        self.axis: np.ndarray = axis
+def stack_nan(coords: np.ndarray) -> np.ndarray:
+    """Stack NaN values to the coordinates array to ensure consistent shape."""
+    stack_array = np.zeros((coords.shape[0], 1, coords.shape[2])) * np.nan
+    return np.concatenate((coords, stack_array), axis=1).reshape(
+        -1, 3, order='C'
+    )
 
-    #    [[x_0, y_0, z_0], #layer 0
-    #     [x_1, y_1, z_1], #layer 1
-    #     ...
-    #    ]
 
-    @property
-    def x_axis(self):
-        return self.axis[:, 0]
+def vectors_to_coords(x, y, z):
+    return np.array([x, y, z]).T
 
-    @property
-    def y_axis(self):
-        return self.axis[:, 1]
 
-    @property
-    def z_axis(self):
-        return self.axis[:, 2]
-
-    @property
-    def angle_rotation(self):
-        # result in degrees
-        trace = np.trace(self.axis, axis1=1, axis2=2)
-        theta = np.arccos((trace - 1) / 2)
-        return np.rad2deg(theta)
-
-    @property
-    def rotation_axis(self) -> np.ndarray:
-        # cf wikipedia
-        # https://en.wikipedia.org/wiki/Rotation_matrix#Conversion from rotation matrix to axis–angle
-        # Do better than this
-        axis_reshaped = np.reshape(
-            np.ravel(self.axis, order="F"), (9, self.axis.shape[0])
-        )
-        rotation_axis = np.array(
-            [
-                axis_reshaped[7] - axis_reshaped[5],
-                axis_reshaped[2] - axis_reshaped[6],
-                axis_reshaped[3] - axis_reshaped[1],
-            ]
-        )
-        rotation_axis_reshaped = np.reshape(
-            np.ravel(rotation_axis, order="F"), (self.axis.shape[0], 3)
-        )
-        simple_truth_table: np.ndarray = np.all(
-            rotation_axis_reshaped == 0, axis=1
-        )
-        is_zero = np.repeat(simple_truth_table[:, np.newaxis], 3, axis=1)
-        default_z_axis = np.array([[0, 0, 1]]).repeat(
-            self.axis.shape[0], axis=0
-        )
-        # creates warning because dividing by zero
-        rotation_axis_normalized = np.where(
-            is_zero,
-            default_z_axis,
-            rotation_axis_reshaped
-            / np.linalg.norm(rotation_axis_reshaped, axis=1)[:, np.newaxis],
-        )
-        return rotation_axis_normalized
-
-    def unflatten(self, flattened_points: np.ndarray):
-        return flattened_points.reshape(self.axis.shape)
-
-    def translate_current(self, translation_vector: np.ndarray):
-        # translation_vector has coords in current frame
-        # rotate translation vector to current frame, using the rotation axis and angle from absolute frame
-        translation_vector_current_frame = rotation_quaternion(
-            translation_vector, -self.angle_rotation, self.rotation_axis
-        )
-        self.origin = self.origin + translation_vector_current_frame
-
-    def translate_absolute(self, translation_vector: np.ndarray):
-        # translation_vector in absolute coords
-        self.origin += translation_vector
-
-    # frame rotation: around itself
-    def rotate(self, line_angles_layer: np.ndarray, rotation_axes: np.ndarray):
-        line_angles_points = np.repeat(
-            line_angles_layer, self.axis.shape[1], axis=0
-        )
-        rotation_axes_array = np.repeat(
-            rotation_axes, self.axis.shape[1], axis=0
-        )
-        flattened_points = self.axis.reshape(
-            -1, 3
-        )  # to refector later with flatten method
-        coords_rotated = rotation_quaternion(
-            flattened_points, line_angles_points, rotation_axes_array
-        )
-        self.axis = self.unflatten(coords_rotated)
-        self.rotation_axis
-
-    def rotate_same_axis(
-        self, line_angles_layer: np.ndarray, rotation_axis: np.ndarray
-    ):
-        line_angle_array = np.repeat(
-            line_angles_layer, self.axis.shape[1], axis=0
-        )
-        flattened_points = self.axis.reshape(-1, 3)
-        coords_rotated = rotation_quaternion_same_axis(
-            flattened_points, line_angle_array, rotation_axis
-        )
-        self.axis = self.unflatten(coords_rotated)
+def layer_to_plot(supports_layers):
+    """Convert the support coordinates to a format suitable for plotting."""
+    return supports_layers.reshape(-1, 3, order='F')
 
 
 class Points:
-    def __init__(self, coords: np.ndarray, frame: Frame = Frame()):
-        self.coords: np.ndarray = coords.copy()  # give copy
-        # [
-        #     [   # layer 0
-        #         [x0_0, y0_0, z0_0],
-        #         [x1_0, y1_0, z1_0]
-        #     ],
-        #     [   # layer 1
-        #         [x0_1, y0_1, z0_1],
-        #         [x1_1, y1_1, z1_1]
-        #     ],
-        #     [   # layer 2
-        #         [x0_2, y0_2, z0_2],
-        #         [x1_2, y1_2, z1_2]
-        #     ],
-        # ]
-        self.frame: Frame = frame  # copy ?
-        # check if coords and frame are compatible
+    """This class handles a set of points in 3D space, represented as a 3D numpy array.
+    The points are stored in a 3D array with shape (number of layers, number of points, 3),
+    where the last dimension represents the x, y, and z coordinates of each point.
 
-    def translate_all(self, translation_vector: np.ndarray):
-        translation_vector_points = np.repeat(
-            translation_vector[:, np.newaxis], self.coords.shape[1], axis=1
+    It provides methods to convert the coordinates to vectors, points, and to create a Points object from.
+    """
+
+    def __init__(self, coords: np.ndarray):
+        if coords.ndim != 3 or coords.shape[2] != 3:
+            raise ValueError(
+                "Coordinates must be a 3D array with shape (number of layers, number of points, 3)"
+            )
+        self.coords = coords
+
+    @property
+    def vectors(self) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+        """Convert the coordinates to vectors."""
+        return (
+            self.coords[:, :, 0].T,
+            self.coords[:, :, 1].T,
+            self.coords[:, :, 2].T,
         )
-        self.coords += translation_vector_points
 
-    def translate_layer(self, translation_vector: np.ndarray, layer: int):
-        self.coords[layer] += translation_vector
+    def points(self, stack=False) -> np.ndarray:
+        """Convert the coordinates to a 2D array of points for plotting or other uses.
+        Returns:
+            np.ndarray: A 2D array of shape (number of points, 3) where each row is a point (x, y, z).
+        """
+        if stack is False:
+            return layer_to_plot(self.coords)
+        else:
+            return stack_nan(self.coords)
 
-    def translate_point(
-        self, translation_vector: np.ndarray, layer: int, index: tuple
-    ):
-        self.coords[layer, index] += translation_vector
+    def flat_layer(self) -> np.ndarray:
+        """Convert the coordinates to a 2D array of points with a column dedicated to layer number for plotting or other uses as dataframe usage.
+        Returns:
+            np.ndarray: A 2D array of shape (number of layers x number of points, 4) where each row is a point (num layer, x, y, z).
+        """
+        raise NotImplementedError
 
-    def flatten(self):
-        return self._flatten(self.coords)
+    def __repr__(self):
+        return f"Points(coords={self.coords})"
+
+    def __len__(self):
+        """Return the number of points."""
+        return self.coords.shape[0]
 
     @staticmethod
-    def _flatten(points):
-        return points.reshape(-1, 3)
+    def from_vectors(x: np.ndarray, y: np.ndarray, z: np.ndarray) -> Self:
+        """Create Points from a vector of coordinates."""
+        if x.ndim != 2 or y.ndim != 2 or z.ndim != 2:
+            raise ValueError("x, y, and z must be 2D arrays")
 
-    def unflatten(self, flattened_points: np.ndarray):
-        return flattened_points.reshape(self.coords.shape)
+        return Points(vectors_to_coords(x, y, z))
 
-    def rotate_all(
-        self, line_angles_layer: np.ndarray, rotation_axes: np.ndarray
-    ):
-        line_angles_points = np.repeat(
-            line_angles_layer, self.coords.shape[1], axis=0
-        )
-        rotation_axes_array = np.repeat(
-            rotation_axes, self.coords.shape[1], axis=0
-        )
-        flattened_points = self.flatten()
-        coords_rotated = rotation_quaternion(
-            flattened_points, line_angles_points, rotation_axes_array
-        )
-        self.coords = self.unflatten(coords_rotated)
+    @staticmethod
+    def from_coords(coords: np.ndarray) -> Self:
+        """Create Points from separate x, y, and z coordinates.
+        Args:
+            coords (np.ndarray): A 3D array of shape (layers, n_points, 3) where each row is a point (x, y, z).
+        """
+        return Points(coords)
 
-    def rotate_one_angle_per_point(
-        self, line_angles_points: np.ndarray, rotation_axes: np.ndarray
-    ):
-        rotation_axes_array = np.repeat(
-            rotation_axes, self.coords.shape[1], axis=0
-        )
-        flattened_points = self.flatten()
-        coords_rotated = rotation_quaternion(
-            flattened_points, line_angles_points, rotation_axes_array
-        )
-        self.coords = self.unflatten(coords_rotated)
 
-    def rotate_same_axis(
-        self, line_angles_layer: np.ndarray, rotation_axis: np.ndarray
-    ):
-        line_angle_array = np.repeat(
-            line_angles_layer, self.coords.shape[1], axis=0
-        )
-        flattened_points = self.flatten()
-        coords_rotated = rotation_quaternion_same_axis(
-            flattened_points, line_angle_array, rotation_axis
-        )
-        self.coords = self.unflatten(coords_rotated)
-
-    def nan_points(self):
-        a, b, c = self.coords.shape
-        return np.hstack([self.coords, np.nan * np.ones((a, 1, c))])
-
-    def coords_for_plot(self):
-        return self._flatten(self.nan_points())
-
-    # transform -> unusable because of order?
-    def transform(
+class SectionPoints:
+    def __init__(
         self,
-        translation_vector: np.ndarray,
+        span_length: np.ndarray,
+        conductor_attachment_altitude: np.ndarray,
+        crossarm_length: np.ndarray,
+        insulator_length: np.ndarray,
         line_angle: np.ndarray,
-        rotation_axis: np.ndarray = np.array([0, 0, 1]),
+        span_model: Span,
+        **_,
     ):
-        self.translate_all(translation_vector)
-        self.rotate_all(line_angle, rotation_axis)
+        (
+            self.supports_ground_coords,
+            self.center_arm_coords,
+            self.arm_coords,
+            self.attachment_coords,
+        ) = get_supports_coords(
+            span_length,
+            line_angle,
+            conductor_attachment_altitude,
+            crossarm_length,
+            insulator_length,
+        )
+        self.plane = CablePlane(
+            span_length,
+            conductor_attachment_altitude,
+            crossarm_length,
+            insulator_length,
+            line_angle,
+        )
 
-    def transform_frame(
+        # self.a = span_length
+        self.line_angle = line_angle
+        # self.b = conductor_attachment_altitude
+        self.crossarm_length = crossarm_length
+        self.insulator_length = insulator_length
+        self._beta = np.array([])
+        self.init_span(span_model)
+
+    def init_span(self, span_model: Span) -> None:
+        self.span_model = span_model
+        self.update_ab()
+        self.set_cable_coordinates(resolution=cfg.graphics.resolution)
+
+    def update_ab(self):
+        self.span_model.span_length = self.plane.a_prime
+        self.span_model.elevation_difference = self.plane.b_prime
+
+    def set_cable_coordinates(self, resolution: int) -> None:
+        self.x_cable: np.ndarray = self.span_model.x(resolution)
+        self.z_cable: np.ndarray = self.span_model.z(self.x_cable)
+
+    @property
+    def beta(self) -> np.ndarray:
+        if self._beta.size == 0:
+            beta = np.zeros(self.x_cable.shape[1])
+        else:
+            beta = self._beta
+        return beta
+
+    @beta.setter
+    def beta(self, value: np.ndarray):
+        if not isinstance(value, np.ndarray):
+            raise TypeError("Beta must be a numpy array")
+        if value.ndim != 1:
+            raise ValueError("Beta must be a 1D array")
+        self._beta = value
+
+    def span_in_cable_frame(self) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+        x_span, y_span, z_span = cable_to_beta_plane(
+            self.x_cable[:, :-1], self.z_cable[:, :-1], beta=self.beta[:-1]
+        )
+        return x_span, y_span, z_span
+
+    def span_in_crossarm_frame(
         self,
-        translation_vector: np.ndarray,
-        line_angle: np.ndarray,
-        rotation_axis: np.ndarray = np.array([0, 0, 1]),
-    ):
-        self.translate_frame(translation_vector)
-        self.rotate_frame(line_angle, rotation_axis)
+    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+        x_span, y_span, z_span = self.span_in_cable_frame()
+        x_span, y_span, z_span = cable_to_crossarm_frame(
+            x_span, y_span, z_span, self.plane.alpha[:-1]
+        )
+        return x_span, y_span, z_span
 
-    def rotate_frame(self, line_angle: np.ndarray, rotation_axis: np.ndarray):
-        self.frame.rotate(line_angle, rotation_axis)
-        self.rotate_all(line_angle, rotation_axis)
+    def span_in_section_frame(
+        self,
+    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
+        x_span, y_span, z_span = self.span_in_crossarm_frame()
+        x_span, y_span, z_span = translate_cable_to_support(
+            x_span,
+            y_span,
+            z_span,
+            self.plane.b,
+            self.plane.a,
+            self.crossarm_length,
+            self.insulator_length,
+            self.line_angle,
+        )
+        return x_span, y_span, z_span
 
-    def translate_frame(self, translation_vector: np.ndarray):
-        self.frame.translate_current(translation_vector)
-        self.translate_all(translation_vector)
+    def get_spans(self, frame) -> Points:
+        if frame == "cable":
+            x_span, y_span, z_span = self.span_in_cable_frame()
+        elif frame == "crossarm":
+            x_span, y_span, z_span = self.span_in_crossarm_frame()
+        elif frame == "section":
+            x_span, y_span, z_span = self.span_in_section_frame()
+        else:
+            raise ValueError("Frame must be 'cable', 'crossarm' or 'section'")
 
-    def coords_given_frame(self, frame):
-        pass
+        return Points.from_vectors(x_span, y_span, z_span)
+
+    def get_supports(self) -> Points:
+        """Get the supports in the global frame."""
+        supports_layers = get_supports_layer(
+            self.supports_ground_coords,
+            self.center_arm_coords,
+            self.arm_coords,
+        )
+        return Points.from_coords(supports_layers)
+
+    def get_insulators(self) -> Points:
+        """Get the insulators in the global frame."""
+        insulator_layers = get_insulator_layer(
+            self.arm_coords,
+            self.attachment_coords,
+        )
+        return Points.from_coords(insulator_layers)
