+++ src/mechaphlowers/core/geometry/references.py
@@ -4,7 +4,7 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 
-from typing import Self, Tuple
+from typing import Tuple
 
 import numpy as np
 
@@ -13,15 +13,16 @@ from mechaphlowers.core.geometry.line_angles import (
     get_attachment_coords,
     get_edge_arm_coords,
     get_elevation_diff_between_supports,
-    get_insulator_layer,
     get_span_lengths_between_supports,
     get_supports_coords,
     get_supports_ground_coords,
-    get_supports_layer,
-    layer_to_plot,
 )
 from mechaphlowers.core.geometry.rotation import rotation_quaternion_same_axis
-from mechaphlowers.core.models.cable.span import Span
+
+""" References for the geometry of the line.
+
+Collections of technical functions to transform coordinates from the different frames of the different objects.
+"""
 
 
 def transform_coordinates(
@@ -242,82 +243,6 @@ def translate_cable_to_support(
     # return x_span, y_span, z_span
 
 
-def vectors_to_coords(x, y, z):
-    return np.array([x, y, z]).T
-
-
-def stack_nan(coords: np.ndarray) -> np.ndarray:
-    """Stack NaN values to the coordinates array to ensure consistent shape."""
-    stack_array = np.zeros((coords.shape[0], 1, coords.shape[2])) * np.nan
-    return np.concatenate((coords, stack_array), axis=1).reshape(
-        -1, 3, order='C'
-    )
-
-
-class Points:
-    """This class handles a set of points in 3D space, represented as a 3D numpy array.
-    The points are stored in a 3D array with shape (number of layers, number of points, 3),
-    where the last dimension represents the x, y, and z coordinates of each point.
-
-    It provides methods to convert the coordinates to vectors, points, and to create a Points object from.
-    """
-
-    def __init__(self, coords: np.ndarray):
-        if coords.ndim != 3 or coords.shape[2] != 3:
-            raise ValueError(
-                "Coordinates must be a 3D array with shape (number of layers, number of points, 3)"
-            )
-        self.coords = coords
-
-    @property
-    def vectors(self) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-        """Convert the coordinates to vectors."""
-        return (
-            self.coords[:, :, 0].T,
-            self.coords[:, :, 1].T,
-            self.coords[:, :, 2].T,
-        )
-
-    def points(self, stack=False) -> np.ndarray:
-        """Convert the coordinates to a 2D array of points for plotting or other uses.
-        Returns:
-            np.ndarray: A 2D array of shape (number of points, 3) where each row is a point (x, y, z).
-        """
-        if stack is False:
-            return layer_to_plot(self.coords)
-        else:
-            return stack_nan(self.coords)
-
-    def flat_layer(self) -> np.ndarray:
-        """Convert the coordinates to a 2D array of points with a column dedicated to layer number for plotting or other uses as dataframe usage.
-        Returns:
-            np.ndarray: A 2D array of shape (number of layers x number of points, 4) where each row is a point (num layer, x, y, z).
-        """
-        raise NotImplementedError
-
-    def __repr__(self):
-        return f"Points(coords={self.coords})"
-
-    def __len__(self):
-        """Return the number of points."""
-        return self.coords.shape[0]
-
-    @staticmethod
-    def from_vectors(x: np.ndarray, y: np.ndarray, z: np.ndarray) -> Self:
-        """Create Points from a vector of coordinates."""
-        if x.ndim != 2 or y.ndim != 2 or z.ndim != 2:
-            raise ValueError("x, y, and z must be 2D arrays")
-
-        return Points(vectors_to_coords(x, y, z))
-
-    @staticmethod
-    def from_coords(coords: np.ndarray) -> Self:
-        """Create Points from separate x, y, and z coordinates.
-        Args:
-            coords (np.ndarray): A 3D array of shape (layers, n_points, 3) where each row is a point (x, y, z).
-        """
-        return Points(coords)
-
 
 class CablePlane:
     """This class handles the parameters for defining the cable plane"""
@@ -340,7 +265,7 @@ class CablePlane:
             line_angle,
             conductor_attachment_altitude,
             crossarm_length,
-            insulator_length
+            insulator_length,
         )
 
         self.a = span_length
@@ -351,137 +276,13 @@ class CablePlane:
         self._beta = np.array([])
 
     @property
-    def a_prime(self):
+    def a_prime(self) -> np.ndarray:
         return get_span_lengths_between_supports(self.attachment_coords)
 
     @property
-    def b_prime(self):
+    def b_prime(self) -> np.ndarray:
         return get_elevation_diff_between_supports(self.attachment_coords)
 
     @property
     def alpha(self) -> np.ndarray:
         return compute_span_azimuth(self.attachment_coords)
-
-
-class SectionPoints:
-    def __init__(
-        self,
-        span_length,
-        conductor_attachment_altitude,
-        crossarm_length,
-        insulator_length,
-        line_angle, span_model, **_
-    ):
-        (
-            self.supports_ground_coords,
-            self.center_arm_coords,
-            self.arm_coords,
-            self.attachment_coords,
-        ) = get_supports_coords(
-            span_length,
-            line_angle,
-            conductor_attachment_altitude,
-            crossarm_length,
-            insulator_length,
-        )
-        self.plane = CablePlane(
-            span_length,
-            conductor_attachment_altitude,
-            crossarm_length,
-            insulator_length,
-            line_angle,
-        )
-        
-        # self.a = span_length
-        self.line_angle = line_angle
-        # self.b = conductor_attachment_altitude
-        self.crossarm_length = crossarm_length
-        self.insulator_length = insulator_length
-        self._beta = np.array([])
-        self.span_model = span_model
-
-    def init_span(self, span_model: Span):
-        self.span_model = span_model
-        self.update_ab()
-        self.set_cable_coordinates()
-
-    def update_ab(self):
-        self.span_model.span_length = self.plane.a_prime
-        self.span_model.elevation_difference = self.plane.b_prime
-
-
-    def set_cable_coordinates(self, resolution: int = 7):
-        self.x_cable: np.ndarray = self.span_model.x(resolution)
-        self.z_cable: np.ndarray = self.span_model.z(self.x_cable)
-
-    @property
-    def beta(self):
-        if self._beta.size == 0:
-            beta = np.zeros(self.x_cable.shape[1])
-        else:
-            beta = self._beta
-        return beta
-
-    @beta.setter
-    def beta(self, value: np.ndarray):
-        if not isinstance(value, np.ndarray):
-            raise TypeError("Beta must be a numpy array")
-        if value.ndim != 1:
-            raise ValueError("Beta must be a 1D array")
-        self._beta = value
-
-    def span_in_cable_frame(self):
-        x_span, y_span, z_span = cable_to_beta_plane(
-            self.x_cable[:, :-1], self.z_cable[:, :-1], beta=self.beta[:-1]
-        )
-        return x_span, y_span, z_span
-
-    def span_in_crossarm_frame(self):
-        x_span, y_span, z_span = self.span_in_cable_frame()
-        x_span, y_span, z_span = cable_to_crossarm_frame(
-            x_span, y_span, z_span, self.plane.alpha[:-1]
-        )
-        return x_span, y_span, z_span
-
-    def span_in_section_frame(self):
-        x_span, y_span, z_span = self.span_in_crossarm_frame()
-        x_span, y_span, z_span = translate_cable_to_support(
-            x_span,
-            y_span,
-            z_span,
-            self.plane.b,
-            self.plane.a,
-            self.crossarm_length,
-            self.insulator_length,
-            self.line_angle,
-        )
-        return x_span, y_span, z_span
-
-    def get_spans(self, frame) -> Points:
-        if frame == "cable":
-            x_span, y_span, z_span = self.span_in_cable_frame()
-        elif frame == "crossarm":
-            x_span, y_span, z_span = self.span_in_crossarm_frame()
-        elif frame == "section":
-            x_span, y_span, z_span = self.span_in_section_frame()
-        else:
-            raise ValueError("Frame must be 'cable', 'crossarm' or 'section'")
-
-        return Points.from_vectors(x_span, y_span, z_span)
-
-    def get_supports(self) -> Points:
-        """Get the supports in the global frame."""
-        supports_layers = get_supports_layer(
-            self.supports_ground_coords,
-            self.center_arm_coords,
-            self.arm_coords,
-        )
-        return Points.from_coords(supports_layers)
-    
-    def get_insulators(self) -> Points:
-        """Get the insulators in the global frame."""
-        insulator_layers = get_insulator_layer(
-            self.arm_coords,
-            self.attachment_coords,
-        )
-        return Points.from_coords(insulator_layers)
