+++ test/core/geometry/test_line_angles.py
@@ -11,55 +11,17 @@ from numpy.testing import assert_allclose
 from pytest import fixture
 
 from mechaphlowers.core.geometry.line_angles import (
+    angle_between_vectors,
     compute_span_azimuth,
-    get_elevation_diff_between_supports,
-    get_attachment_coords,
     get_edge_arm_coords,
+    get_elevation_diff_between_supports,
     get_span_lengths_between_supports,
     get_supports_coords,
     get_supports_ground_coords,
     get_supports_layer,
-    layer_to_plot,
 )
-from mechaphlowers.core.models.cable.span import CatenarySpan
+from mechaphlowers.core.geometry.points import layer_to_plot
 from mechaphlowers.entities.arrays import SectionArray
-from mechaphlowers.core.geometry.references import (
-    SectionPoints,
-    cable_to_beta_plane,
-    cable_to_crossarm_frame,
-    spans_to_vector,
-    translate_cable_to_support,
-)
-from mechaphlowers.plotting.plot import set_layout
-
-
-def plot_points_3d(fig, points):
-    fig.add_trace(
-        go.Scatter3d(
-            x=points[:, 0],
-            y=points[:, 1],
-            z=points[:, 2],
-            mode='markers+lines',
-            marker=dict(
-                size=3,
-            ),  # color='red'),
-            name='Points',
-        )
-    )
-
-
-def plot_points_2d(fig, points):
-    fig.add_trace(
-        go.Scatter(
-            x=points[:, 0],
-            y=points[:, 1],
-            mode='markers+lines',
-            marker=dict(
-                size=4,
-            ),  # color='red'),
-            name='Points',
-        )
-    )
 
 
 @fixture
@@ -82,7 +44,7 @@ def section_array_line_angles():
     return section_array
 
 
-def test_get_supports_coords(section_array_line_angles):
+def test_get_supports_ground_coords(section_array_line_angles):
     span_length = section_array_line_angles.data.span_length.to_numpy()
     line_angle = section_array_line_angles.data.line_angle.to_numpy()
     supports_ground_coords = get_supports_ground_coords(
@@ -98,11 +60,11 @@ def test_get_supports_coords(section_array_line_angles):
     )
     np.testing.assert_allclose(supports_ground_coords, expected_coords)
 
-    fig = go.Figure()
-    plot_points_2d(fig, supports_ground_coords)
-
-
-# fig.show()
+    # import plotly.graph_objects as go
+    # from mechaphlowers.core.geometry.points import plot_points_2d
+    # fig = go.Figure()
+    # plot_points_2d(fig, supports_ground_coords)
+    # fig.show()
 
 
 def test_build_supports(section_array_line_angles):
@@ -134,25 +96,23 @@ def test_build_supports(section_array_line_angles):
 
     np.testing.assert_allclose(arm_coords, expected_coords)
 
+    # import plotly.graph_objects as go
+    # from mechaphlowers.core.geometry.points import plot_points_3d
     # fig = go.Figure()
     # plot_points_3d(fig, supports_ground_coords)
     # plot_points_3d(fig, center_arm_coords)
     # plot_points_3d(fig, arm_coords)
-
     # fig.show()
 
 
 def test_get_supports(section_array_line_angles):
-    """Get the supports in the global frame."""
+    """Get the supports in the support frame."""
     span_length = section_array_line_angles.data.span_length.to_numpy()
     line_angle = section_array_line_angles.data.line_angle.to_numpy()
     conductor_attachment_altitude = (
         section_array_line_angles.data.conductor_attachment_altitude.to_numpy()
     )
     crossarm_length = section_array_line_angles.data.crossarm_length.to_numpy()
-    insulator_length = (
-        section_array_line_angles.data.insulator_length.to_numpy()
-    )
 
     supports_ground_coords = get_supports_ground_coords(
         span_length, line_angle
@@ -163,49 +123,32 @@ def test_get_supports(section_array_line_angles):
         line_angle,
         crossarm_length,
     )
-    attachment_coords = get_attachment_coords(
-        edge_arm_coords, insulator_length
-    )
 
     supports_layer = get_supports_layer(
         supports_ground_coords,
         center_arm_coords,
         edge_arm_coords,
-        attachment_coords,
     )
     supports_coords = layer_to_plot(supports_layer)
 
     expected_coords = np.array(
         [
             [0.0, 0.0, 0.0],
-            [0.0, 0.0, 30.0],
-            [0.0, 40.0, 30.0],
-            [0.0, 40.0, 30.0],
-            [np.nan, np.nan, np.nan],
             [500.0, 0.0, 0.0],
-            [500.0, 0.0, 40.0],
-            [507.65366865, 18.47759065, 40.0],
-            [507.65366865, 18.47759065, 35.0],
-            [np.nan, np.nan, np.nan],
             [825.26911935, -325.26911935, 0.0],
-            [825.26911935, -325.26911935, 60.0],
-            [817.50454799, -354.24689413, 60.0],
-            [817.50454799, -354.24689413, 50.0],
-            [np.nan, np.nan, np.nan],
             [1327.55054902, -190.68321589, 0.0],
+            [0.0, 0.0, 30.0],
+            [500.0, 0.0, 40.0],
+            [825.26911935, -325.26911935, 60.0],
             [1327.55054902, -190.68321589, 70.0],
+            [0.0, 40.0, 30.0],
+            [507.65366865, 18.47759065, 40.0],
+            [817.50454799, -354.24689413, 60.0],
             [1327.55054902, -240.68321589, 70.0],
-            [1327.55054902, -240.68321589, 70.0],
-            [np.nan, np.nan, np.nan],
         ]
     )
 
     assert_allclose(supports_coords, expected_coords)
-
-    # fig = go.Figure()
-    # plot_points_3d(fig, supports_coords)
-    # # set_layout(fig)
-    # fig.show()
 
 
 def test_span_lengths(section_array_line_angles):
@@ -257,222 +200,114 @@ def test_get_supports_coords(section_array_line_angles):
     assert True
 
 
-
-
-
-
-
-def test_span_absolute_coords_new_obj(section_array_line_angles):
-
-
-    span_model = CatenarySpan(**section_array_line_angles.to_numpy())
-    
-    s = SectionPoints(
-        span_model=span_model, **section_array_line_angles.to_numpy()
-    )
-    
-    
-
-    s.init_span(span_model)
-
-    fig = go.Figure()
-    plot_points_3d(fig, s.get_spans("cable").points(True))
-    plot_points_3d(fig, s.get_spans("crossarm").points(True))
-    plot_points_3d(fig, s.get_spans("section").points(True))
-
-    plot_points_3d(fig, s.get_supports().points(True))
-    plot_points_3d(fig, s.get_insulators().points(True))
-
-    set_layout(fig)
-    fig.show()
-    assert True
-
-
-def test_span_absolute_coords(section_array_line_angles):
-    # ---- Same code than previous test ----""
-    span_length = section_array_line_angles.data.span_length.to_numpy()
-    line_angle = section_array_line_angles.data.line_angle.to_numpy()
-    conductor_attachment_altitude = (
-        section_array_line_angles.data.conductor_attachment_altitude.to_numpy()
-    )
-    crossarm_length = section_array_line_angles.data.crossarm_length.to_numpy()
-    insulator_length = (
-        section_array_line_angles.data.insulator_length.to_numpy()
+def test_compute_span_azimuth():
+    attachement_coords = np.array(
+        [
+            [0.0, 40.0, 30.0],
+            [507.65366865, 18.47759065, 35.0],
+            [817.50454799, -354.24689413, -22.0],
+            [1327.55054902, -240.68321589, 70.0],
+        ]
     )
 
-    (
-        supports_ground_coords,
-        center_arm_coords,
-        arm_coords,
-        attachment_coords,
-    ) = get_supports_coords(
-        span_length,
-        line_angle,
-        conductor_attachment_altitude,
-        crossarm_length,
-        insulator_length,
+    expected_azimuth = np.array(
+        [-2.42764954, -50.26279675, 12.55236927, 11.93819593]
     )
 
-    new_span_length = get_span_lengths_between_supports(attachment_coords)
-    new_elevation_diff = get_elevation_diff_between_supports(attachment_coords)
-    # ----------
-
-    span_model = CatenarySpan(**section_array_line_angles.to_numpy())
-    span_model.span_length = new_span_length
-    span_model.elevation_difference = new_elevation_diff
-
-    x = span_model.x(21)
-    z = span_model.z(x)
-
-    pts = spans_to_vector(x, x * 0, z)
-
-    fig = go.Figure()
-    # plot_points_3d(fig, pts)
-    # fig.show()
-
-    # get pts for spans
-
-    beta = np.array([50.0, 10.0, 0.0, np.nan])
-    x_span, y_span, z_span = cable_to_beta_plane(
-        x[:, :-1], z[:, :-1], beta=beta[:-1]
+    np.testing.assert_allclose(
+        compute_span_azimuth(attachement_coords), expected_azimuth
     )
 
-    pts = spans_to_vector(x_span, y_span, z_span)
 
-    # just to verify cable frame
-    plot_points_3d(fig, pts)
+def test_angle_between_vectors():
+    # Define input vectors
+    vector_a = np.array([[1, 0], [0, 1]]).T
+    vector_b = np.array([[0, 1], [1, 0]]).T
 
-    alpha = compute_span_azimuth(attachment_coords)
-    x_span, y_span, z_span = cable_to_crossarm_frame(
-        x_span, y_span, z_span, alpha[:-1]
-    )
+    # Call the function
+    result = angle_between_vectors(vector_a, vector_b)
 
-    pts = spans_to_vector(x_span, y_span, z_span)
-    # just to verify spans in crossarm frame
-    plot_points_3d(fig, pts)
+    # Define the expected result
+    expected_result = np.array([np.pi / 2, -np.pi / 2])
 
-    x_span, y_span, z_span = translate_cable_to_support(
-        x_span,
-        y_span,
-        z_span,
-        conductor_attachment_altitude,
-        span_length,
-        crossarm_length,
-        insulator_length,
-        line_angle,
-    )
-
-    pts = spans_to_vector(x_span, y_span, z_span)
-    # final span plot
-    plot_points_3d(fig, pts)
-
-    supports_layer = get_supports_layer(
-        supports_ground_coords,
-        center_arm_coords,
-        arm_coords,
-        attachment_coords,
-    )
-    supports_coords = layer_to_plot(supports_layer)
-    plot_points_3d(fig, supports_coords)
-
-    set_layout(fig)
-
-    fig.show()
-    fig.to_html()
-
-    assert True
+    # Assert that the result matches the expected result
+    np.testing.assert_array_almost_equal(result, expected_result, decimal=5)
 
 
-def test_span_absolute_coords_(section_array_line_angles):
-    # ---- Same code than previous test ----""
-    span_length = section_array_line_angles.data.span_length.to_numpy()
-    line_angle = section_array_line_angles.data.line_angle.to_numpy()
-    conductor_attachment_altitude = (
-        section_array_line_angles.data.conductor_attachment_altitude.to_numpy()
-    )
-    crossarm_length = section_array_line_angles.data.crossarm_length.to_numpy()
-    insulator_length = (
-        section_array_line_angles.data.insulator_length.to_numpy()
-    )
+def test_angle_between_vectors_parallel():
+    # Define input vectors that are parallel
+    vector_a = np.array([[1, 0], [0, 1]]).T
+    vector_b = np.array([[2, 0], [0, 2]]).T
 
-    (
-        supports_ground_coords,
-        center_arm_coords,
-        arm_coords,
-        attachment_coords,
-    ) = get_supports_coords(
-        span_length,
-        line_angle,
-        conductor_attachment_altitude,
-        crossarm_length,
-        insulator_length,
-    )
+    # Call the function
+    result = angle_between_vectors(vector_a, vector_b)
 
-    new_span_length = get_span_lengths_between_supports(attachment_coords)
-    new_elevation_diff = get_elevation_diff_between_supports(attachment_coords)
-    # ----------
+    # Define the expected result
+    expected_result = np.array([0, 0])
 
-    span_model = CatenarySpan(**section_array_line_angles.to_numpy())
-    span_model.span_length = new_span_length
-    span_model.elevation_difference = new_elevation_diff
+    # Assert that the result matches the expected result
+    np.testing.assert_array_almost_equal(result, expected_result, decimal=5)
 
-    x = span_model.x(21)
-    z = span_model.z(x)
 
-    pts = spans_to_vector(x, x * 0, z)
+def test_angle_between_vectors_antiparallel():
+    # Define input vectors that are antiparallel
+    vector_a = np.array(
+        [
+            [
+                1,
+                0,
+            ],
+            [
+                0,
+                1,
+            ],
+        ]
+    ).T
+    vector_b = np.array(
+        [
+            [
+                -1,
+                0,
+            ],
+            [
+                0,
+                -1,
+            ],
+        ]
+    ).T
 
-    fig = go.Figure()
-    # plot_points_3d(fig, pts)
-    # fig.show()
+    # Call the function
+    result = angle_between_vectors(vector_a, vector_b)
 
-    # get pts for spans
+    # Define the expected result
+    expected_result = np.array([np.pi, np.pi])
 
-    beta = np.array([50.0, 10.0, 0.0, np.nan])
-    x_span, y_span, z_span = cable_to_beta_plane(
-        x[:, :-1], z[:, :-1], beta=beta[:-1]
-    )
+    # Assert that the result matches the expected result
+    np.testing.assert_array_almost_equal(result, expected_result, decimal=5)
 
-    pts = spans_to_vector(x_span, y_span, z_span)
 
-    # just to verify cable frame
-    plot_points_3d(fig, pts)
+def test_angle_between_vectors_zero_vector():
+    # Define input vectors with a zero vector
+    vector_a = np.array([[1, 0, 0], [0, 1, 0]]).T
+    vector_b = np.array([[0, 0, 0], [0, 0, 0]]).T
 
-    alpha = compute_span_azimuth(attachment_coords)
-    x_span, y_span, z_span = cable_to_crossarm_frame(
-        x_span, y_span, z_span, alpha[:-1]
-    )
+    # Call the function
+    result = angle_between_vectors(vector_a, vector_b)
+    expected_result = np.array([np.nan, np.nan, np.nan])
 
-    pts = spans_to_vector(x_span, y_span, z_span)
-    # just to verify spans in crossarm frame
-    plot_points_3d(fig, pts)
+    np.testing.assert_array_almost_equal(result, expected_result, decimal=5)
 
-    x_span, y_span, z_span = translate_cable_to_support(
-        x_span,
-        y_span,
-        z_span,
-        conductor_attachment_altitude,
-        span_length,
-        crossarm_length,
-        insulator_length,
-        line_angle,
-    )
 
-    pts = spans_to_vector(x_span, y_span, z_span)
-    # final span plot
-    plot_points_3d(fig, pts)
+def test_angle_between_vectors_same_vector():
+    # Define input vectors that are the same
+    vector_a = np.array([[1, 0, 0], [0, 1, 0]]).T
+    vector_b = np.array([[1, 0, 0], [0, 1, 0]]).T
 
-    supports_layer = get_supports_layer(
-        supports_ground_coords,
-        center_arm_coords,
-        arm_coords,
-        attachment_coords,
-    )
-    supports_coords = layer_to_plot(supports_layer)
-    plot_points_3d(fig, supports_coords)
+    # Call the function
+    result = angle_between_vectors(vector_a, vector_b)
 
-    set_layout(fig)
+    # Define the expected result
+    expected_result = np.array([0, 0, np.nan])
 
-    fig.show()
-    fig.to_html()
-
-    assert True
+    # Assert that the result matches the expected result
+    np.testing.assert_array_almost_equal(result, expected_result, decimal=5)
