+++ test/test_config.py
@@ -19,7 +19,7 @@ def test_config_on_plot(default_section_array_three_spans):
     assert (
         fig._data[0].get('marker').get('size') == options.graphics.marker_size
     )
-    assert fig._data[0].get('x').shape[0] == options.graphics.resolution * (
+    assert fig._data[0].get('x').shape[0] == (options.graphics.resolution+1) * (
         default_section_array_three_spans.data.shape[0] - 1
     )
     # fig.show() # deactivate for auto unit testing
